Changed: Total refactor using internal millis()
In this implementation, we're only using the RTC to initialize millisOffset.  After that, the RTC isn't used.  In updateBTParts(), we're relying on millis(), in a similar fashion as we did in our first prototype without an RTC.  Because we're consistently using millis() as the time authority, and it has the proper resolution, we are no longer "missing seconds."